+++ src/components/DescriptionList/Description.tsx
@@ -1,10 +1,17 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import { Col } from 'antd';
+import { ColProps } from 'antd/es/col';
+import React from 'react';
 import styles from './index.less';
 import responsive from './responsive';
 
-const Description = ({ term, column, children, ...restProps }) => (
+export interface IDescriptionProps extends ColProps {
+  column?: number;
+  key?: string | number;
+  style?: React.CSSProperties;
+  term?: React.ReactNode;
+}
+
+const Description: React.SFC<IDescriptionProps> = ({ term, column, children, ...restProps }) => (
   <Col {...responsive[column]} {...restProps}>
     {term && <div className={styles.term}>{term}</div>}
     {children !== null && children !== undefined && <div className={styles.detail}>{children}</div>}
@@ -13,10 +20,6 @@ const Description = ({ term, column, children, ...restProps }) => (
 
 Description.defaultProps = {
   term: '',
-};
-
-Description.propTypes = {
-  term: PropTypes.node,
 };
 
 export default Description;
