--- src/components/DescriptionList/index.js
@@ -1,5 +1,0 @@
-import DescriptionList from './DescriptionList';
-import Description from './Description';
-
-DescriptionList.Description = Description;
-export default DescriptionList;
