--- src/components/DescriptionList/responsive.js
@@ -1,6 +1,0 @@
-export default {
-  1: { xs: 24 },
-  2: { xs: 24, sm: 12 },
-  3: { xs: 24, sm: 12, md: 8 },
-  4: { xs: 24, sm: 12, md: 6 },
-};
