+++ src/components/DescriptionList/responsive.ts
@@ -1,0 +1,10 @@
+import { ColProps } from 'antd/es/col';
+
+const responsive: { [key: number]: ColProps } = {
+  1: { xs: 24 },
+  2: { xs: 24, sm: 12 },
+  3: { xs: 24, sm: 12, md: 8 },
+  4: { xs: 24, sm: 12, md: 6 },
+};
+
+export default responsive;
